+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Add optional `schema` argument for `SqlFetcher`.
+- Improve error reporting for unmapped required placeholders; warn about potential override issues.
 
 ## [0.2.1] - 2023-02-04